refactor(metrics): moved to lib folder
